@@ -1,85 +1,72 @@
 {
-  "name": "material-app",
-  "version": "2.0.1",
-  "description": "Material App Pro - React Admin & Dashboard Template",
-  "author": "Paul Laros <support@bootlab.io>",
-  "private": true,
-  "license": "UNLICENSED",
+  "name": "react-webpack-setting",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject",
-    "lint": "eslint src --ext .js,.jsx,.ts,.tsx --fix"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 10",
-    "not op_mini all"
-  ],
-  "dependencies": {
-    "@date-io/core": "1.3.13",
-    "@date-io/date-fns": "1.3.13",
-    "@fullcalendar/core": "5.5.0",
-    "@fullcalendar/daygrid": "5.5.0",
-    "@fullcalendar/interaction": "5.5.0",
-    "@fullcalendar/react": "5.5.0",
-    "@material-ui/core": "4.11.2",
-    "@material-ui/data-grid": "4.0.0-alpha.13",
-    "@material-ui/docs": "4.0.0-beta.3",
-    "@material-ui/icons": "4.11.2",
-    "@material-ui/lab": "4.0.0-alpha.57",
-    "@material-ui/pickers": "3.2.10",
-    "@material-ui/styles": "4.11.2",
-    "@material-ui/system": "4.11.2",
-    "axios": "0.21.1",
-    "axios-mock-adapter": "1.19.0",
-    "bootstrap": "5.1.0",
-    "chart.js": "2.9.4",
-    "css-vendor": "2.0.8",
-    "date-fns": "2.16.1",
-    "deepmerge": "4.2.2",
-    "formik": "2.2.6",
-    "google-map-react": "2.1.9",
-    "immutable": "^4.0.0",
-    "jss": "10.5.0",
-    "material-ui-dropzone": "3.5.0",
-    "material-ui-search-bar": "^1.0.0",
-    "polished": "4.0.5",
-    "query-string": "^7.0.1",
-    "react": "^17.0.1",
-    "react-app-polyfill": "2.0.0",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-bootstrap": "^2.0.0-beta.6",
-    "react-chartjs-2": "2.11.1",
-    "react-cookie": "^4.0.3",
-    "react-daum-postcode": "^3.0.1",
-    "react-dom": "^17.0.1",
-    "react-dragula": "1.1.17",
-    "react-feather": "2.0.9",
-    "react-helmet": "6.1.0",
-    "react-js-pagination": "^3.0.3",
-    "react-jvectormap": "0.0.16",
-    "react-material-ui-form-validator": "^2.1.4",
-    "react-perfect-scrollbar": "1.5.8",
-    "react-phone-input-2": "^2.14.0",
-    "react-quill": "1.3.5",
-    "react-redux": "7.2.2",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.1",
-    "react-syntax-highlighter": "15.4.3",
-    "redux": "4.0.5",
-    "redux-actions": "^2.6.5",
-    "redux-thunk": "2.3.0",
-    "styled-components": "5.2.1",
-    "yup": "0.32.8"
+    "dev": "webpack serve --env development",
+    "build": "webpack --progress"
   },
   "devDependencies": {
-    "eslint-config-prettier": "6.15.0",
-    "eslint-plugin-prettier": "3.3.0",
-    "prettier": "2.2.1",
-    "prop-types": "15.7.2",
-    "redux-devtools-extension": "2.13.8"
+    "@babel/core": "^7.13.8",
+    "@babel/preset-env": "^7.13.9",
+    "@babel/preset-react": "^7.12.13",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^6.5.1",
+    "eslint": "^7.2.0",
+    "eslint-config-airbnb": "18.2.1",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.3.1",
+    "mini-css-extract-plugin": "^1.3.9",
+    "prettier": "^2.2.1",
+    "redux-devtools-extension": "^2.13.9",
+    "webpack": "^5.36.1",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2"
   },
-  "proxy": "http://localhost:8088"
+  "dependencies": {
+    "@babel/plugin-transform-runtime": "^7.16.5",
+    "@babel/runtime": "^7.16.5",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.4",
+    "@mui/lab": "^5.0.0-alpha.61",
+    "@mui/material": "^5.2.4",
+    "@mui/styles": "^5.2.3",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
+    "axios": "^0.24.0",
+    "dotenv-webpack": "^7.0.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "formik": "^2.2.9",
+    "query-string": "^7.1.0",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.0",
+    "react-daum-postcode": "^3.0.1",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^12.0.4",
+    "react-helmet": "^6.1.0",
+    "react-helmet-async": "^1.2.2",
+    "react-js-pagination": "^3.0.3",
+    "react-query": "^3.39.2",
+    "react-redux": "^7.2.6",
+    "react-refresh": "^0.14.0",
+    "react-router-dom": "^6.4.5",
+    "redux": "^4.1.2",
+    "redux-actions": "^2.6.5",
+    "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0",
+    "redux-saga": "^1.1.3",
+    "style-loader": "^3.3.1",
+    "styled-components": "^5.3.3",
+    "use-react-router": "^1.0.7",
+    "yup": "^0.32.11"
+  }
 }
