@@ -1,135 +1,3 @@
-// Theme
-export const SUCCESS_CODE = 0;
-export const FAIL_CODE = -1;
-
-export const THEME_SET = "THEME_SET";
-export const THEMES = {
-  DEFAULT: "DEFAULT",
-  DARK: "DARK",
-  LIGHT: "LIGHT",
-  BLUE: "BLUE",
-  GREEN: "GREEN",
-  INDIGO: "INDIGO",
-};
-
-//common
-export const FTP_URL = "https://storage.googleapis.com/careercompany";
-export const DATA_REQUEST = "DATA_REQUEST";
-export const DATA_SUCCESS = "DATA_SUCCESS";
-export const DATA_FAILURE = "DATA_FAILURE";
-
-// Auth
-export const ROLE_MEMBER = "ROLE_MEMBER";
-export const ROLE_ADMIN = "ROLE_ADMIN";
-
-export const VALIDATE_TOKEN_REQUEST = "VALIDATE_TOKEN_REQUEST";
-export const VALIDATE_TOKEN_SUCCESS = "VALIDATE_TOKEN_SUCCESS";
-export const VALIDATE_TOKEN_FAILURE = "VALIDATE_TOKEN_FAILURE";
-
-export const REISSUE_TOKEN_REQUEST = "REISSUE_TOKEN_REQUEST";
-export const REISSUE_TOKEN_SUCCESS = "REISSUE_TOKEN_SUCCESS";
-export const REISSUE_TOKEN_FAILURE = "REISSUE_TOKEN_FAILURE";
-
-export const AUTH_SIGN_UP_REQUEST = "AUTH_SIGN_UP_REQUEST";
-export const AUTH_SIGN_UP_SUCCESS = "AUTH_SIGN_UP_SUCCESS";
-export const AUTH_SIGN_UP_FAILURE = "AUTH_SIGN_UP_FAILURE";
-
-export const AUTH_LOGOUT = "AUTH_LOGOUT";
-
-export const AUTH_GET_INFO_REQUEST = "AUTH_GET_INFO_REQUEST";
-export const AUTH_GET_INFO_SUCCESS = "AUTH_GET_INFO_SUCCESS";
-export const AUTH_GET_INFO_FAILURE = "AUTH_GET_INFO_FAILURE";
-
-
-export const AUTH_CHECK_ID_REQUEST = "AUTH_CHECK_ID_REQUEST";
-export const AUTH_CHECK_ID_SUCCESS = "AUTH_CHECK_ID_SUCCESS";
-export const AUTH_CHECK_ID_FAILURE = "AUTH_CHECK_ID_FAILURE";
-
-export const AUTH_FIND_ID_REQUEST = "AUTH_FIND_ID_REQUEST";
-export const AUTH_FIND_ID_SUCCESS = "AUTH_FIND_ID_SUCCESS";
-export const AUTH_FIND_ID_FAILURE = "AUTH_FIND_ID_FAILURE";
-
-export const AUTH_SMS_VALIDATE_REQUEST = "AUTH_SMS_VALIDATE_REQUEST";
-export const AUTH_SMS_VALIDATE_SUCCESS = "AUTH_SMS_VALIDATE_SUCCESS";
-export const AUTH_SMS_VALIDATE_FAILURE = "AUTH_SMS_VALIDATE_FAILURE";
-
-
-export const AUTH_LOGIN_REQUEST = "AUTH_LOGIN_REQUEST";
-export const AUTH_LOGIN_SUCCESS = "AUTH_LOGIN_SUCCESS";
-export const AUTH_LOGIN_FAILURE = "AUTH_LOGIN_FAILURE";
-export const AUTH_SIGN_OUT = "AUTH_SIGN_OUT";
-
-export const AUTH_RESET_PASSWORD_REQUEST = "AUTH_RESET_PASSWORD_REQUEST";
-export const AUTH_RESET_PASSWORD_SUCCESS = "AUTH_RESET_PASSWORD_SUCCESS";
-export const AUTH_RESET_PASSWORD_FAILURE = "AUTH_RESET_PASSWORD_FAILURE";
-
-//inpection
-export const INSPECTION_LIST_REQUEST = "INSPECTION_LIST_REQUEST";
-export const INSPECTION_LIST_SUCCESS = "INSPECTION_LIST_SUCCESS";
-export const INSPECTION_LIST_FAILURE = "INSPECTION_LIST_FAILURE";
-
-//question
-export const QUESTION_TYPE_TEXT = "TEXT";
-export const QUESTION_TYPE_IMAGE = "IMAGE";
-
-//result
-export const RESULT_LIST_REQUEST = "RESULT_LIST_REQUEST";
-export const RESULT_LIST_SUCCESS = "RESULT_LIST_SUCCESS";
-export const RESULT_LIST_FAILURE = "RESULT_LIST_FAILURE";
-
-// Member
-export const MEMBER_LIST_REQUEST = "MEMBER_LIST_REQUEST";
-export const MEMBER_LIST_SUCCESS = "MEMBER_LIST_SUCCESS";
-export const MEMBER_LIST_FAILURE = "MEMBER_LIST_FAILURE";
-
-export const MEMBER_RESET = "MEMBER_RESET";
-export const MEMBER_UPDATE_REQUEST = "MEMBER_UPDATE_REQUEST";
-export const MEMBER_UPDATE_SUCCESS = "MEMBER_UPDATE_SUCCESS";
-export const MEMBER_UPDATE_FAILURE = "MEMBER_UPDATE_FAILURE";
-
-//Group
-export const CLEAR_GROUP = "CLEAR_GROUP";
-export const GROUP_LIST_REQUEST = "GROUP_LIST_REQUEST";
-export const GROUP_LIST_SUCCESS = "GROUP_LIST_SUCCESS";
-export const GROUP_LIST_FAILURE = "GROUP_LIST_FAILURE";
-
-export const GROUP_DETAIL_REQUEST = "GROUP_DETAIL_REQUEST";
-export const GROUP_DETAIL_SUCCESS = "GROUP_DETAIL_SUCCESS";
-export const GROUP_DETAIL_FAILURE = "GROUP_DETAIL_FAILURE";
-
-export const GROUP_REGIST_REQUEST = "GROUP_REGIST_REQUEST";
-export const GROUP_REGIST_SUCCESS = "GROUP_REGIST_SUCCESS";
-export const GROUP_REGIST_FAILURE = "GROUP_REGIST_FAILURE";
-
-export const GROUP_UPDATE_REQUEST = "GROUP_UPDATE_REQUEST";
-export const GROUP_UPDATE_SUCCESS = "GROUP_UPDATE_SUCCESS";
-export const GROUP_UPDATE_FAILURE = "GROUP_UPDATE_FAILURE";
-//user
-
-export const USER_LIST_REQUEST = "USER_LIST_REQUEST";
-export const USER_LIST_SUCCESS = "USER_LIST_SUCCESS";
-export const USER_LIST_FAILURE = "USER_LIST_FAILURE";
-
-export const USER_DELETE_REQUEST = "USER_DELETE_REQUEST";
-export const USER_DELETE_SUCCESS = "USER_DELETE_SUCCESS";
-export const USER_DELETE_FAILURE = "USER_DELETE_FAILURE"
-
-export const USER_MODIFY_REQUEST = "USER_MODIFY_REQUEST";
-export const USER_MODIFY_SUCCESS = "USER_MODIFY_SUCCESS";
-export const USER_MODIFY_FAILURE = "USER_MODIFY_FAILURE"
-
-
-export const USER_ANSWER_LIST_REQUEST = "USER_ANSWER_LIST_REQUEST";
-export const USER_ANSWER_LIST_SUCCESS = "USER_ANSWER_LIST_SUCCESS";
-export const USER_ANSWER_LIST_FAILURE = "USER_ANSWER_LIST_FAILURE";
-
-export const USER_ANSWER_REGIST_REQUEST = "USER_ANSWER_REGIST_REQUEST";
-export const USER_ANSWER_REGIST_SUCCESS = "USER_ANSWER_REGIST_SUCCESS";
-export const USER_ANSWER_REGIST_FAILURE = "USER_ANSWER_REGIST_FAILURE";
-
-export const USER_RESET = "USER_RESET";
-
-//SMS
-export const SMS_SEND_REQUEST = "SMS_SEND_REQUEST";
-export const SMS_SEND_SUCCESS = "SMS_SEND_SUCCESS";
-export const SMS_SEND_FAILURE = "SMS_SEND_FAILURE";
+export const FTP_URL = 'https://storage.googleapis.com/careercompany';
+export const TYPE_TEXT = 'TEXT';
+export const TYPE_IMAGE = 'IMAGE'
